Revert "Refac for gtests"
