+++ CMakeLists.txt
@@ -3,8 +3,6 @@ cmake_minimum_required(VERSION 3.5)
 project(ExampleProj)
 
 include_directories(src/functions)
-include_directories(src/window)
-include_directories(src/helpers)
 
 set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
