+++ src/functions/sum_function.cpp
@@ -1,4 +1,3 @@
-#include "sum_function.h"
 
 int sum(int x, int y)
 {
