--- src/helpers/CMakeLists.txt
@@ -1,3 +1,0 @@
-project(helpers)
-
-add_library(helpers STATIC helpers.cpp helpers.h)
