--- src/window/CMakeLists.txt
@@ -1,3 +1,0 @@
-project(window)
-
-add_library(window STATIC window.cpp window.h)
